+++ components/collection-card/CollectionCard_style.module.css
@@ -1,0 +1,26 @@
+.collection__card {
+    width: 320px;
+    height: 100%;
+    text-align: center;
+}
+
+.man {
+    width: 100%;
+    height: 430px;
+    object-fit: cover;
+}
+
+.collection__card p {
+    padding: 30px 10px;
+    font-weight: 500;
+    font-size: 16px;
+    color: #414141;
+}
+
+@media (max-width: 500px) {
+    .collection__card {
+        width: 100%;
+        height: 100%;
+        text-align: center;
+    }
+}
