+++ components/collection/Collection-style.module.css
@@ -19,25 +19,6 @@
     column-gap: 20px;
 }
 
-.collection__card {
-    width: 320px;
-    height: 100%;
-    text-align: center;
-}
-
-.man {
-    width: 100%;
-    height: 430px;
-    object-fit: cover;
-}
-
-.collection__card p {
-    padding: 30px 10px;
-    font-weight: 500;
-    font-size: 16px;
-    color: #414141;
-}
-
 .collection__slider {
     display: none;
 }
@@ -72,12 +53,6 @@
     .collection__slider {
         display: block;
         width: 100%;
-    }
-
-    .collection__card {
-        width: 100%;
-        height: 100%;
-        text-align: center;
     }
 }
