+++ src/pages/index.js
@@ -2,7 +2,7 @@ import Head from 'next/head'
 import Header from '../../components/header/Header'
 import First from '../../components/first/First'
 import Brands from '../../components/brands/Brands'
-import Collection from '../../components/colllection/Collection'
+import Collection from '../../components/collection/Collection'
 import Novelties from '../../components/novelties/Novelties'
 import Perfumes from '../../components/perfumes/Perfumes'
 import Promotion from '../../components/promotion/Promotion'
